Added confirmation to reset button

+++ src/reset/reset.ts
@@ -19,7 +19,55 @@ export async function reset(interaction: ChatInputCommandInteraction<CacheType>,
         await db.userGame.delete({where: {id: userGames?.id}})
         return await interaction.reply({ephemeral: true, content: `Successfully deleted ${game} from your play history.`})
     } else {
-        await db.userGame.deleteMany({where: {userId: user?.id}})
-        return await interaction.reply({ephemeral: true, content: "All game history has been deleted!"})
+        await interaction.reply({ephemeral: true, content: `Are you sure you want to delete all of your play history? This cannot be undone.`})
+        // Add a confirmation button here
+        const filter = (i: any) => i.user.id === interaction.user.id;
+        const collector = interaction.channel?.createMessageComponentCollector({ filter,
+            time: 15000
+        });
+        collector?.on('collect', async (i: any) => {
+            if (i.customId === 'confirm') {
+                await db.userGame.deleteMany({where: {userId: user?.id}})
+                await i.update({content: `Successfully deleted all of your play history.`, components: []})
+            } else if (i.customId === 'cancel') {
+                await i.update({content: `Cancelled deletion of all of your play history.`, components: []})
+            }
+        });
+
+        collector?.on('end', async (collected, reason) => {
+            if (reason === 'time') {
+                await interaction.editReply({content: `Timed out.`, components: []})
+            }
+        });
+
+        return await interaction.editReply({ content: `Are you sure you want to delete all of your play history? This cannot be undone.`, components: [
+            {
+                type: 1,
+                components: [
+                    {
+                        type: 2,
+                        style: 3,
+                        label: 'Confirm',
+                        customId: 'confirm'
+                    },
+                    {
+                        type: 2,
+                        style: 4,
+                        label: 'Cancel',
+                        customId: 'cancel'
+                    }
+                ]
+            }
+        ]})
+        // const filter = (i: any) => i.user.id === interaction.user.id;
+        // const collector = interaction.channel?.createMessageComponentCollector({ filter, time: 15000 });
+        // collector?.on('collect', async i => {
+        //     if (i.customId === 'yes') {
+        //         await db.userGame.deleteMany({where: {userId: user?.id}})
+        //         await i.update({content: `Successfully deleted all of your play history.`, components: []})
+        //     } else if (i.customId === 'no') {
+        //         await i.update({content: `Cancelled.`, components: []})
+        //     }
+        // });
     }
 }
